deployment/inference/code/requirements.txt
https://github.com/er-ri/bertjsc/archive/refs/tags/v1.0.0.tar.gz
transformers
fugashi
ipadic
sentencepiece
pytorch-lightning